@@ -8,10 +8,11 @@ Plas::Application.configure do
 
   # Log error messages when you accidentally call methods on nil.
   config.whiny_nils = true
-
+  config.assets.enabled = false
+  config.assets.compress = false
+  config.serve_static_assets = true
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
-  config.action_view.debug_rjs             = true
   config.action_controller.perform_caching = false
 
   # Don't care if the mailer can't send
@@ -31,5 +32,8 @@ Plas::Application.configure do
     config.logger = Logger.new(Rails.root.join("log",Rails.env + ".log"),3,5*1024*1024)  
   end
   config.autoload_paths += Dir["#{config.root}/lib/**/"] # include all subdirectories
+
+  require "#{config.root}/lib/rack/firebug_logger.rb"
+#  config.middleware.use ::Rack::FirebugLogger
 
 end
