@@ -1,14 +1,17 @@
 Plas::Application.configure do
-  # Settings specified here will take precedence over those in config/environment.rb
-	#deprecation errors to stderr
-  config.active_support.deprecation = :stderr
+  # Settings specified here will take precedence over those in config/application.rb
+
   # The test environment is used exclusively to run your application's
   # test suite.  You never need to work with it otherwise.  Remember that
   # your test database is "scratch space" for the test suite and is wiped
   # and recreated between test runs.  Don't rely on the data there!
   config.cache_classes = true
 
-  # Log error messages when you accidentally call methods on nil.
+  # Configure static asset server for tests with Cache-Control for performance
+  config.serve_static_assets = true
+  config.static_cache_control = "public, max-age=3600"
+
+  # Log error messages when you accidentally call methods on nil
   config.whiny_nils = true
 
   # Show full error reports and disable caching
@@ -30,14 +33,10 @@ Plas::Application.configure do
   # This is necessary if your schema can't be completely dumped by the schema dumper,
   # like if you have constraints or database-specific column types
   # config.active_record.schema_format = :sql
-  config.after_initialize do
-    #set activemerchant to use the gateways' test servers
-    ActiveMerchant::Billing::Base.mode = :test
-  end
-  if $0 == "irb"
-    config.logger = Logger.new(STDOUT)
-  else
-    config.logger = Logger.new(Rails.root.join("log",Rails.env + ".log"),3,5*1024*1024)  
-  end
 
+  # Print deprecation notices to the stderr
+  config.active_support.deprecation = :stderr
+
+  # Allow pass debug_assets=true as a query parameter to load pages with unpackaged assets
+  config.assets.allow_debugging = true
 end
