@@ -2,6 +2,6 @@
 
 # Your secret key for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
-# Make sure the secret is at least 30 characters and all random, 
+# Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Rails.application.config.secret_token = '3946c687ee75110b1897f1c69a70a8fb9ea257ba271b1856637fc56a1fed4243661612002e32ce2418b6ba4e3d9249760cdfbedf0ab20912756abc69a6d61cb1'
+Plas::Application.config.secret_token = 'f60af8d1eb8578e47af8b8421de342f9a1fcc9c3adb04db36d1c9db1bf856ba22a50b29e5dc18f6bc089bbf2f632e9050f33ee9ea2d0685f331312b8a87893a4'
